organised parts of the code into namespaces
